+++ app/hidden-browser-window-preload.js
@@ -1,16 +1,49 @@
-// Allow jQuery to run without Node collisions
-// http://electron.atom.io/docs/v0.37.2/faq/electron-faq/#i-can-not-use-jqueryrequirejsmeteorangularjs-in-electron
-window.nodeRequire = require;
-delete window.require;
-delete window.exports;
-delete window.module;
+/*!
+ * =============================================================================
+ * Internal abilities required for Markbot to function
+ * Trying to isolate all the Node.js functionality away from the user scripts for security
+ * =============================================================================
+ */
+window.__markbot = (function () {
+  'use strict';
 
-window.__markbotGetBrowserWindow = function () {
-  return nodeRequire('electron').remote.BrowserWindow.fromId(window.__markbot_hidden_browser_window_id).webContents;
-};
+  const getCurrentTaskWindowId = function () {
+    if (window.__markbotHiddenTestingWindowId) return window.__markbotHiddenTestingWindowId;
 
+    return document.referrer.replace(/^https:\/\//, '').replace(/\.running-task-windows\.markbot\.web\-dev\.tools\/$/, '');
+  };
+
+  const sendMessageToWindow = function (windowId, messageId, ...message) {
+    if (typeof windowId === 'string') windowId = parseInt(windowId, 10);
+
+    require('electron').remote.BrowserWindow.fromId(windowId).webContents.send(messageId, ...message);
+  };
+
+  const sendInputEventToWindow = function (windowId, inputEvent) {
+    if (typeof windowId === 'string') windowId = parseInt(windowId, 10);
+
+    require('electron').remote.BrowserWindow.fromId(windowId).webContents.sendInputEvent(inputEvent);
+  };
+
+  const getPerformanceTestingService = function () {
+    return require('webcoach');
+  };
+
+  return {
+    getCurrentTaskWindowId: getCurrentTaskWindowId,
+    getPerformanceTestingService: getPerformanceTestingService,
+    sendMessageToWindow: sendMessageToWindow,
+    sendInputEventToWindow: sendInputEventToWindow,
+  };
+}());
+
+/*!
+ * =============================================================================
+ * Document & window event catchers for events Markbot needs to know about
+ * =============================================================================
+ */
 window.addEventListener('error', function (err) {
-  __markbotGetBrowserWindow().send('__markbot-functionality-error', err.message, err.lineno, err.filename);
+  window.__markbot.sendMessageToWindow(window.__markbot.getCurrentTaskWindowId(), '__markbot-functionality-error', err.message, err.lineno, err.filename);
 });
 
 window.addEventListener('load', function (ev) {
@@ -19,7 +52,7 @@ window.addEventListener('load', function (ev) {
       window.requestAnimationFrame(function () {
         window.requestAnimationFrame(function () {
           process.nextTick(function () {
-            window.__markbotGetBrowserWindow().send('__markbot-hidden-browser-window-loaded', {location: window.location.href});
+            window.__markbot.sendMessageToWindow(window.__markbot.getCurrentTaskWindowId(), '__markbot-hidden-browser-window-loaded', {location: window.location.href});
           });
         });
       });
@@ -33,7 +66,7 @@ document.fonts.ready.then(function () {
       window.requestAnimationFrame(function () {
         window.requestAnimationFrame(function () {
           process.nextTick(function () {
-            window.__markbotGetBrowserWindow().send('__markbot-hidden-browser-window-fonts-loaded', {location: window.location.href});
+            window.__markbot.sendMessageToWindow(window.__markbot.getCurrentTaskWindowId(), '__markbot-hidden-browser-window-fonts-loaded', {location: window.location.href});
           });
         });
       });
@@ -41,6 +74,11 @@ document.fonts.ready.then(function () {
   });
 });
 
+/*!
+ * =============================================================================
+ * Overwrite some Javascript functions so they can be better tested
+ * =============================================================================
+ */
 window.alert = function (str) {
   return true;
 };